+++ src/app.html
@@ -35,7 +35,8 @@
 
 		/* Bg colour and image */
 		background-color: #243347;
-		background-image: url('/bg10.png');
+		background-image: url('/background-image.svg');
+		background-attachment: fixed;
 		background-size: cover;
 		background-position: center;
 	}
