+++ src/routes/welcome/signup/credentials/+page.svelte
@@ -6,27 +6,7 @@
 	import BoxComponent from '../../../../components/Box/Box_Component.svelte';
 </script>
 
-<body>
-	<div class="frame">
-		<HeaderComponent text="Almost there..." />
-		<BoxComponent boxClass="signup-box">
-			<UserCredentialsComponent />
-		</BoxComponent>
-	</div>
-</body>
-
-<style>
-	.frame {
-		height: 100vh; 
-		width: 100vw;
-		background-color: #243347;
-		background-image: url('/bg10.png');
-		background-size: cover;
-		background-position: center;
-
-		/* I know flexboxes are dodgy but trust me ;) */
-		display: flex;
-		flex-direction: column;
-		flex-wrap: nowrap;
-	}
-</style>
+<HeaderComponent text="Almost there..." />
+<BoxComponent boxClass="signup-box">
+	<UserCredentialsComponent />
+</BoxComponent>
